Add files via upload
Everything works on my end. The only feature I couldn't get working was for the memos in the collection to display on hover. I could only get EVERY memo to display on hover, but couldn't get the correct memos from the specific collection to display. I feel like you could do it by using a <form> to grab the correct collection_id and memo_id's but then the page refreshes and all is lost. Spent many hours trying to get this to work. The main hurdle is just grabbing the right collection_id on a hover without using a <form>.

ALSO ADDED REPLIES TABLE IN SQL FILE